use server side fetching for lockedSMS
