+++ packages/lib/server/repository/event.ts
@@ -1,5 +1,5 @@
 import { getPublicEvent } from "@calcom/features/eventtypes/lib/getPublicEvent";
-import prisma from "@calcom/prisma";
+import { prisma } from "@calcom/prisma";
 
 import type { TEventInputSchema } from "../../../trpc/server/routers/publicViewer/event.schema";
