+++ packages/lib/server/repository/featureFlag.ts
@@ -1,4 +1,4 @@
-import prisma from "@calcom/prisma";
+import { prisma } from "@calcom/prisma";
 
 export class FeatureFlagRepository {
   static async getFeatureFlags() {
