+++ packages/lib/server/repository/schedule.ts
@@ -1,4 +1,4 @@
-import prisma from "@calcom/prisma";
+import { prisma } from "@calcom/prisma";
 import { getDefaultScheduleId } from "@calcom/trpc/server/routers/viewer/availability/util";
 
 import { hasReadPermissionsForUserId } from "../../hasEditPermissionForUser";
